requirements.txt
GitPython==3.1.45
PyYAML==6.0.3
requests==2.32.5
